feat offer book

+++ lib/data/data_providers/remote/remote_user_source.dart
@@ -3,7 +3,9 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:share_take/constants/api.dart';
 import 'package:share_take/data/models/book/book_local.dart';
+import 'package:share_take/data/models/book_offers/book_offer_remote.dart';
 import 'package:share_take/data/models/book_wants/book_wants_remote.dart';
+import 'package:share_take/data/models/request/book_offer_request.dart';
 import 'package:share_take/data/models/request/book_want_request.dart';
 import 'package:share_take/data/models/request/register_request.dart';
 import 'package:share_take/data/models/user/user_local.dart';
@@ -139,8 +141,70 @@ class RemoteUserSource {
       String currentUserId = firebaseAuth.currentUser!.uid;
 
       await fireStore
-        .collection(StaticApi.wantedCollection)
-        .doc(currentUserId + bookId).delete();
+          .collection(StaticApi.wantedCollection)
+          .doc(currentUserId + bookId).delete();
+    } on FirebaseException catch (firebaseException) {
+      throw Exception(firebaseException.message);
+    }
+  }
+
+
+  //OFFER LIST
+  Future<List<BookOfferRemote>> getUserOfferList(String userId) async {
+    try {
+      CollectionReference offeredCollection = fireStore.collection(StaticApi.offeredCollection);
+      List<QueryDocumentSnapshot> userOfferedItemList =
+      await offeredCollection.where('userId', isEqualTo: userId).get().then((snapshot) => snapshot.docs);
+      if (userOfferedItemList.isNotEmpty) {
+        return userOfferedItemList.map((e) => BookOfferRemote.fromSnapshot(e)).toList();
+      } else {
+        return [];
+      }
+    } on FirebaseException catch (firebaseException) {
+      throw Exception(firebaseException.message);
+    }
+  }
+
+  Future addBookToOfferList(
+      String bookId,
+      ) async {
+    if (!_userLoggedIn()) {
+      throw Exception("User not found. Please login");
+    }
+
+    try {
+      String currentUserId = firebaseAuth.currentUser!.uid;
+      List<BookOfferRemote> offers = await getUserOfferList(currentUserId);
+
+      BookOfferRemote? found;
+
+      try {
+        found = offers.firstWhere((element) => element.bookId == bookId);
+      } catch (e) {}
+
+      if (found != null) {
+        throw Exception("You are already offering this book");
+      } else {
+        await fireStore
+            .collection(StaticApi.offeredCollection)
+            .doc(currentUserId + bookId)
+            .set(BookOfferRequest(userId: currentUserId, bookId: bookId).toMap());
+      }
+    } on FirebaseException catch (firebaseException) {
+      throw Exception(firebaseException.message);
+    }
+  }
+
+  Future removeBookOffer(String bookId) async {
+    if (!_userLoggedIn()) {
+      throw Exception("User not found. Please login");
+    }
+    try {
+      String currentUserId = firebaseAuth.currentUser!.uid;
+
+      await fireStore
+          .collection(StaticApi.offeredCollection)
+          .doc(currentUserId + bookId).delete();
     } on FirebaseException catch (firebaseException) {
       throw Exception(firebaseException.message);
     }
